init push
Contains the new definition that extends the vanilla walls. This should allow any and all code changes to walls to automatically apply to this mod as well.

Drastically reduces lines of codes from ~250 to 12.

Fixes the issue of not being able to build on bridges and other more niche wall functionality (such as meditation) not applying to this mod's walls.

Old code is included in another directory. That directory should not be uploaded to steam.